Disallow fileops on values
