+++ src/shell/shell.rs
@@ -1,12 +1,37 @@
-use crate::commands::command::EvaluatedWholeStreamCommandArgs;
+use crate::commands::command::{EvaluatedWholeStreamCommandArgs, RunnablePerItemContext};
+use crate::commands::cp::CopyArgs;
+use crate::commands::mkdir::MkdirArgs;
+use crate::commands::mv::MoveArgs;
+use crate::commands::rm::RemoveArgs;
 use crate::context::SourceMap;
 use crate::errors::ShellError;
+use crate::prelude::*;
 use crate::stream::OutputStream;
 
 pub trait Shell: std::fmt::Debug {
     fn name(&self, source_map: &SourceMap) -> String;
     fn ls(&self, args: EvaluatedWholeStreamCommandArgs) -> Result<OutputStream, ShellError>;
     fn cd(&self, args: EvaluatedWholeStreamCommandArgs) -> Result<OutputStream, ShellError>;
+    fn cp(
+        &self,
+        args: CopyArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError>;
+    fn mkdir(
+        &self,
+        args: MkdirArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError>;
+    fn mv(
+        &self,
+        args: MoveArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError>;
+    fn rm(
+        &self,
+        args: RemoveArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError>;
     fn path(&self) -> String;
     fn set_path(&mut self, path: String);
