+++ src/shell/value_shell.rs
@@ -1,4 +1,8 @@
-use crate::commands::command::EvaluatedWholeStreamCommandArgs;
+use crate::commands::command::{EvaluatedWholeStreamCommandArgs, RunnablePerItemContext};
+use crate::commands::cp::CopyArgs;
+use crate::commands::mkdir::MkdirArgs;
+use crate::commands::mv::MoveArgs;
+use crate::commands::rm::RemoveArgs;
 use crate::context::SourceMap;
 use crate::prelude::*;
 use crate::shell::shell::Shell;
@@ -97,6 +101,54 @@ impl Shell for ValueShell {
         let mut stream = VecDeque::new();
         stream.push_back(ReturnSuccess::change_cwd(path));
         Ok(stream.into())
+    }
+
+    fn cp(
+        &self,
+        _args: CopyArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError> {
+        Err(ShellError::labeled_error(
+            "cp not currently supported on values",
+            "not currently supported",
+            context.name,
+        ))
+    }
+
+    fn mv(
+        &self,
+        _args: MoveArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError> {
+        Err(ShellError::labeled_error(
+            "mv not currently supported on values",
+            "not currently supported",
+            context.name,
+        ))
+    }
+
+    fn mkdir(
+        &self,
+        _args: MkdirArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError> {
+        Err(ShellError::labeled_error(
+            "mkdir not currently supported on values",
+            "not currently supported",
+            context.name,
+        ))
+    }
+
+    fn rm(
+        &self,
+        _args: RemoveArgs,
+        context: &RunnablePerItemContext,
+    ) -> Result<VecDeque<ReturnValue>, ShellError> {
+        Err(ShellError::labeled_error(
+            "rm not currently supported on values",
+            "not currently supported",
+            context.name,
+        ))
     }
 
     fn path(&self) -> String {
